migrations, seeders and models

+++ database/migrations/0001_01_01_000000_create_users_table.php
@@ -4,20 +4,22 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
-    public function up(): void
-    {
+    public function up(): void {
         Schema::create('users', function (Blueprint $table) {
-            $table->id();
+            $table->id()->autoIncrement();
             $table->string('name');
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
+            $table->string('phone')->unique();
+            $table->string('description')->nullable();
+            $table->tinyInteger('active', unsigned: true);
             $table->rememberToken();
+            $table->softDeletes();
             $table->timestamps();
         });
 
@@ -40,9 +42,10 @@ return new class extends Migration
     /**
      * Reverse the migrations.
      */
-    public function down(): void
-    {
-        Schema::dropIfExists('users');
+    public function down(): void {
+        if (!Schema::hasTable('academy_owners')) {
+            Schema::dropIfExists('users');
+        }
         Schema::dropIfExists('password_reset_tokens');
         Schema::dropIfExists('sessions');
     }
